\name{CNew-package}
\alias{CNew-package}
\alias{CNew}
\docType{package}
\title{
\packageTitle{CNew}
}
\description{
\packageDescription{CNew}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CNew}
\packageIndices{CNew}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CNew}

Maintainer: \packageMaintainer{CNew}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
